receiving_app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
SQLAlchemy==2.0.34
pyodbc==5.1.0
pandas==2.2.2
openpyxl==3.1.5
reportlab==4.0.7
python-multipart==0.0.6
pillow==10.1.0
qrcode==7.4.2
